vk: mass move vulkan-domain files to vulkan/
Just plain move, no structural changes.
