+++ ref/vk/vk_core.c
@@ -3,26 +3,25 @@
 #include "vk_common.h"
 #include "r_textures.h"
 #include "vk_overlay.h"
-#include "vk_image.h"
-#include "vk_staging.h"
+#include "vulkan/VImage.h"
+#include "vulkan/VStaging.h"
 #include "vk_framectl.h"
 #include "vk_brush.h"
 #include "vk_scene.h"
 #include "vk_cvar.h"
-#include "vk_pipeline.h"
+#include "vulkan/VPipeline.h"
 #include "vk_render.h"
 #include "vk_geometry.h"
 #include "vk_studio.h"
 #include "vk_rtx.h"
-#include "vk_descriptor.h"
-#include "vk_resources.h"
-#include "vk_nv_aftermath.h"
-#include "vk_devmem.h"
-#include "vk_commandpool.h"
+#include "vulkan/VDescriptor.h"
+#include "vulkan/VResource.h"
+#include "vulkan/VNvAftermath.h"
+#include "vulkan/VDevmem.h"
 #include "r_speeds.h"
 #include "vk_sprite.h"
 #include "vk_beams.h"
-#include "vk_combuf.h"
+#include "vulkan/VCombuf.h"
 #include "vk_entity_data.h"
 #include "vk_logs.h"
 #include "std/arrays.h"
