+++ ref/vk/vk_framectl.c
@@ -4,15 +4,15 @@
 #include "vk_scene.h"
 #include "vk_render.h"
 #include "vk_cvar.h"
-#include "vk_devmem.h"
-#include "vk_swapchain.h"
-#include "vk_image.h"
-#include "vk_staging.h"
-#include "vk_commandpool.h"
-#include "vk_combuf.h"
+#include "vulkan/VDevmem.h"
+#include "vulkan/VSwapchain.h"
+#include "vulkan/VImage.h"
+#include "vulkan/VStaging.h"
+#include "vulkan/VCommandPool.h"
+#include "vulkan/VCombuf.h"
 #include "vk_logs.h"
-#include "vk_barrier.h"
-#include "vk_resources.h"
+#include "vulkan/VBarrier.h"
+#include "vulkan/VResource.h"
 
 #include "std/arrays.h"
 #include "std/profiler.h"
