+++ ref/vk/vk_light.c
@@ -1,5 +1,5 @@
 #include "vk_light.h"
-#include "vk_buffer.h"
+#include "vulkan/VBuffer.h"
 #include "vk_mapents.h"
 #include "r_textures.h"
 #include "vk_lightmap.h"
@@ -7,11 +7,11 @@
 #include "shaders/ray_interop.h"
 #include "std/bitarray.h"
 #include "std/profiler.h"
-#include "vk_staging.h"
+#include "vulkan/VStaging.h"
 #include "r_speeds.h"
 #include "vk_logs.h"
 #include "vk_framectl.h"
-#include "vk_resources.h"
+#include "vulkan/VResource.h"
 
 #include "mod_local.h"
 #include "xash3d_mathlib.h"
