+++ ref/vk/vk_render.c
@@ -1,19 +1,19 @@
 #include "vk_render.h"
 
 #include "vk_core.h"
-#include "vk_buffer.h"
+#include "vulkan/VBuffer.h"
 #include "vk_geometry.h"
-#include "vk_barrier.h"
-#include "vk_resources.h"
-#include "vk_combuf.h"
+#include "vulkan/VBarrier.h"
+#include "vulkan/VResource.h"
+#include "vulkan/VCombuf.h"
 #include "vk_const.h"
 #include "vk_common.h"
 #include "vk_cvar.h"
-#include "vk_pipeline.h"
+#include "vulkan/VPipeline.h"
 #include "vk_textures.h"
 #include "vk_math.h"
 #include "vk_rtx.h"
-#include "vk_descriptor.h"
+#include "vulkan/VDescriptor.h"
 #include "std/alolcator.h"
 #include "std/profiler.h"
 #include "r_speeds.h"
