+++ ref/vk/vk_textures.h
@@ -2,7 +2,7 @@
 
 #include "r_textures.h"
 #include "vk_core.h"
-#include "vk_image.h"
+#include "vulkan/VImage.h"
 #include "vk_const.h"
 
 #include "std/unordered_roadmap.h"
