+++ ref/vk/vulkan/VCombuf.c
@@ -1,5 +1,5 @@
-#include "vk_combuf.h"
-#include "vk_commandpool.h"
+#include "VCombuf.h"
+#include "VCommandPool.h"
 #include "vk_logs.h"
 
 #include "std/profiler.h"
