+++ ref/vk/vulkan/VImageExtra.h
