+++ ref/vk/vulkan/VMeatpipe.c
@@ -1,9 +1,9 @@
-#include "vk_meatpipe.h"
+#include "VMeatpipe.h"
 
-#include "vk_pipeline.h"
-#include "vk_resources.h"
-#include "vk_barrier.h"
-#include "ray_pass.h"
+#include "VPipeline.h"
+#include "VResource.h"
+#include "VBarrier.h"
+#include "VPass.h"
 #include "vk_common.h"
 #include "vk_logs.h"
 #include "std/profiler.h"
