+++ ref/vk/vulkan/VMisc.c
