+++ ref/vk/vulkan/VRayAccel.c
@@ -6,13 +6,13 @@
 #include "vk_rtx.h"
 #include "vk_ray_internal.h"
 #include "r_speeds.h"
-#include "vk_combuf.h"
-#include "vk_barrier.h"
+#include "VCombuf.h"
+#include "VBarrier.h"
 #include "vk_math.h"
 #include "vk_geometry.h"
 #include "vk_render.h"
 #include "vk_logs.h"
-#include "vk_resources.h"
+#include "vulkan/VResource.h"
 
 #include "std/arrays.h"
 #include "std/profiler.h"
