+++ ref/vk/wscript
@@ -91,7 +91,7 @@ def build(bld):
 	defines = []
 	libpath = []
 
-	source = bld.path.ant_glob(['*.c', 'std/*.c'])
+	source = bld.path.ant_glob(['*.c', 'std/*.c', 'vulkan/*.c'])
 	glsl_source = bld.path.ant_glob(['shaders/*.vert', 'shaders/*.frag'])
 	rtx_glsl_source = bld.path.ant_glob(['shaders/*.rgen', 'shaders/*.rchit', 'shaders/*.rmiss', 'shaders/*.rahit', 'shaders/*.comp'])
